conda-requirements.txt
python=3.8
google-api-python-client=1.8
tqdm=4.46.0
pylint=2.5.0
google-auth-oauthlib=0.4.1
flask=1.1.2
pandas=1.0.3
apscheduler=3.6.3
sqlalchemy=1.3.17
python-dotenv=0.13.0
flask-login=0.5.0
pip=20.1
gspread=3.6.0=pypi_0
